requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.9
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.0
